Refactor objects resource into smaller modules
Split resource.server.ts into:
- schemas.ts: Zod schemas and types
- cache.server.ts: Count caching logic
- history.server.ts: History recording
- actions.server.ts: Action handlers
- resource.server.ts: Thin orchestrator
